deeptech/requirements.txt
asgiref==3.4.1
attrs==21.2.0
coverage==5.5
Django==3.2.6
django-debug-toolbar==3.2.1
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
importlib-metadata==4.6.3
iniconfig==1.1.1
packaging==21.0
phonenumbers==8.12.28
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
zipp==3.5.0
